Make stat calculation capable of returning a much broader range of stats.
Fix level calculation (was log(linear), should have been log(square))

Add in mean and standard deviation, and a host of potential tile levels.

This will break auto-tiling for the moment; once we figure out which potential level determination is the right one, I'll re-adjust to make it work agian.